Fixed a few warnings.

+++ src/main.c
@@ -409,10 +409,10 @@ int dbm_main(int argc, char **argv)
    */
 #if STACK_DIRECTION < 0
   /* Equvivalent with |= 0xffff */
-  Pike_interpreter.stack_top += (~((unsigned long)Pike_interpreter.stack_top)) & 0xffff;
+  Pike_interpreter.stack_top += (~((size_t)Pike_interpreter.stack_top)) & 0xffff;
 #else /* STACK_DIRECTION >= 0 */
   /* Equvivalent with &= ~0xffff */
-  Pike_interpreter.stack_top -= ( ((unsigned long)Pike_interpreter.stack_top)) & 0xffff;
+  Pike_interpreter.stack_top -= ( ((size_t)Pike_interpreter.stack_top)) & 0xffff;
 #endif /* STACK_DIRECTION < 0 */
 
 #if defined(HAVE_GETRLIMIT) && defined(RLIMIT_STACK)
